pyproject.toml
[project]
name = "odoo-oca-odoorpc-integration-ex"
version = "0.0.3"
description = "odoo-oca-odoorpc-integration-ex"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "odoorpc>=0.10.1,<2",
]
